fix test report collect bug

+++ entry.sh
@@ -214,12 +214,15 @@ if [ ${ACTION} == "test" ]; then
       if [ ! -z "$test_done" ]; then
         echo "Test status: test done"
         if [ ! -z "$is_mvn_cmd" ]; then
-          echo "Copy test reports"
-          mkdir -p test_report
-          cd test_report
-          kubectl cp test-${ns}:/root/code/${TEST_CODE_PATH}/target/surefire-reports/. . -n ${ns}
-          rm -rf *.txt
-          ls
+          if [ ! -d "./test_report" ]; then
+            echo "Copy test reports"
+            mkdir -p test_report
+            cd test_report
+            kubectl cp test-${ns}:/root/code/${TEST_CODE_PATH}/target/surefire-reports/. . -n ${ns}
+            rm -rf *.txt
+            ls
+            cd -
+          fi
         fi
       fi
   done
